# agent-platform/acceptance/round2/_openapi.json
{"openapi":"3.0.2","info":{"title":"Agent 人机协作平台","version":"1.0.0"},"paths":{"/api/login":{"post":{"tags":["auth"],"summary":"Login","operationId":"login_api_login_post","requestBody":{"content":{"application/json":{"schema":{"$ref":"#/components/schemas/LoginIn"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/me":{"get":{"tags":["auth"],"summary":"Me","operationId":"me_api_me_get","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/org/tree":{"get":{"tags":["org"],"summary":"Org Tree","description":"5 平台嵌套部门，部门含 people[] 与 agents[]（基本信息）","operationId":"org_tree_api_org_tree_get","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/people":{"get":{"tags":["org"],"summary":"List People","operationId":"list_people_api_people_get","parameters":[{"required":false,"schema":{"title":"Tier","type":"string"},"name":"tier","in":"query"},{"required":false,"schema":{"title":"Dept Id","type":"integer"},"name":"dept_id","in":"query"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/people/{pid}":{"get":{"tags":["org"],"summary":"Get Person","operationId":"get_person_api_people__pid__get","parameters":[{"required":true,"schema":{"title":"Pid","type":"integer"},"name":"pid","in":"path"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/agents":{"get":{"tags":["agents"],"summary":"List Agents","operationId":"list_agents_api_agents_get","parameters":[{"required":false,"schema":{"title":"Platform Id","type":"integer"},"name":"platform_id","in":"query"},{"required":false,"schema":{"title":"Status","type":"string"},"name":"status","in":"query"},{"required":false,"schema":{"title":"Wave","type":"integer"},"name":"wave","in":"query"},{"required":false,"schema":{"title":"Category","type":"string"},"name":"category","in":"query"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/agents/{aid}":{"get":{"tags":["agents"],"summary":"Get Agent","operationId":"get_agent_api_agents__aid__get","parameters":[{"required":true,"schema":{"title":"Aid","type":"integer"},"name":"aid","in":"path"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"patch":{"tags":["agents"],"summary":"Update Agent","description":"状态维护：仅 boss/coach 或该数字员工的 owner 本人；status 有枚举校验。","operationId":"update_agent_api_agents__aid__patch","parameters":[{"required":true,"schema":{"title":"Aid","type":"integer"},"name":"aid","in":"path"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"title":"Body","type":"object"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/scenarios":{"get":{"tags":["scenarios"],"summary":"List Scenarios","operationId":"list_scenarios_api_scenarios_get","parameters":[{"required":false,"schema":{"title":"Platform Id","type":"integer"},"name":"platform_id","in":"query"},{"required":false,"schema":{"title":"Dept Id","type":"integer"},"name":"dept_id","in":"query"},{"required":false,"schema":{"title":"Status","type":"string"},"name":"status","in":"query"},{"required":false,"schema":{"title":"Priority","type":"string"},"name":"priority","in":"query"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["scenarios"],"summary":"Create Scenario","description":"敏捷立项申报：创建即为\"待立项\"","operationId":"create_scenario_api_scenarios_post","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"title":"Body","type":"object"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/scenarios/{sid}/initiate":{"post":{"tags":["scenarios"],"summary":"Initiate Scenario","description":"立项：状态→已立项，自动建项目工作区并拉入关联 agent 与申请人","operationId":"initiate_scenario_api_scenarios__sid__initiate_post","parameters":[{"required":true,"schema":{"title":"Sid","type":"integer"},"name":"sid","in":"path"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workspaces":{"get":{"tags":["workspaces"],"summary":"List Workspaces","operationId":"list_workspaces_api_workspaces_get","parameters":[{"required":false,"schema":{"title":"Type","type":"string"},"name":"type","in":"query"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["workspaces"],"summary":"Create Workspace","operationId":"create_workspace_api_workspaces_post","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"title":"Body","type":"object"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workspaces/{wid}":{"get":{"tags":["workspaces"],"summary":"Get Workspace","operationId":"get_workspace_api_workspaces__wid__get","parameters":[{"required":true,"schema":{"title":"Wid","type":"integer"},"name":"wid","in":"path"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/workspaces/{wid}/messages":{"get":{"tags":["workspaces"],"summary":"List Messages","operationId":"list_messages_api_workspaces__wid__messages_get","parameters":[{"required":true,"schema":{"title":"Wid","type":"integer"},"name":"wid","in":"path"},{"required":false,"schema":{"title":"Zone","type":"string"},"name":"zone","in":"query"},{"required":false,"schema":{"title":"Limit","type":"integer","default":200},"name":"limit","in":"query"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["workspaces"],"summary":"Post Message","description":"发言；zone=='agent' 或内容 @数字员工 时触发数字员工执行","operationId":"post_message_api_workspaces__wid__messages_post","parameters":[{"required":true,"schema":{"title":"Wid","type":"integer"},"name":"wid","in":"path"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"title":"Body","type":"object"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tasks":{"get":{"tags":["tasks"],"summary":"List Tasks","operationId":"list_tasks_api_tasks_get","parameters":[{"required":false,"schema":{"title":"Status","type":"string"},"name":"status","in":"query"},{"required":false,"schema":{"title":"Agent Id","type":"integer"},"name":"agent_id","in":"query"},{"required":false,"schema":{"title":"Reviewer Id","type":"integer"},"name":"reviewer_id","in":"query"},{"required":false,"schema":{"title":"Workspace Id","type":"integer"},"name":"workspace_id","in":"query"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["tasks"],"summary":"Create Task","description":"任务中心直建任务。\n\n- 带 agent_id：复用引擎执行逻辑，立即产出交付物并转「待审核」；\n- 不带 agent_id：允许创建（停留「待处理」），响应附带 hint 提示不会自动执行。","operationId":"create_task_api_tasks_post","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"title":"Body","type":"object"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/tasks/{tid}/review":{"post":{"tags":["tasks"],"summary":"Review Task","description":"人工审核：approve 通过计产出 / reject 驳回并触发数字员工重做。\n\n权限：仅 tier ∈ {boss, coach, backbone}，且不能审核自己发起的任务。","operationId":"review_task_api_tasks__tid__review_post","parameters":[{"required":true,"schema":{"title":"Tid","type":"integer"},"name":"tid","in":"path"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"title":"Body","type":"object"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/skills":{"get":{"tags":["skills"],"summary":"List Skills","operationId":"list_skills_api_skills_get","parameters":[{"required":false,"schema":{"title":"Scope","type":"string"},"name":"scope","in":"query"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["skills"],"summary":"Create Skill","operationId":"create_skill_api_skills_post","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"title":"Body","type":"object"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/knowledge/spaces":{"get":{"tags":["knowledge"],"summary":"List Spaces","operationId":"list_spaces_api_knowledge_spaces_get","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/knowledge/documents":{"get":{"tags":["knowledge"],"summary":"List Documents","operationId":"list_documents_api_knowledge_documents_get","parameters":[{"required":false,"schema":{"title":"Space Id","type":"integer"},"name":"space_id","in":"query"},{"required":false,"schema":{"title":"Level","type":"string"},"name":"level","in":"query"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["knowledge"],"summary":"Create Document","operationId":"create_document_api_knowledge_documents_post","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"title":"Body","type":"object"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/metrics/dashboard":{"get":{"tags":["metrics"],"summary":"Dashboard","operationId":"dashboard_api_metrics_dashboard_get","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/metrics/agents":{"get":{"tags":["metrics"],"summary":"Agent Metrics","operationId":"agent_metrics_api_metrics_agents_get","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/governance/incentives":{"get":{"tags":["governance"],"summary":"List Incentives","operationId":"list_incentives_api_governance_incentives_get","parameters":[{"required":false,"schema":{"title":"Status","type":"string"},"name":"status","in":"query"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["governance"],"summary":"Create Incentive","operationId":"create_incentive_api_governance_incentives_post","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"title":"Body","type":"object"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/governance/reimbursements":{"get":{"tags":["governance"],"summary":"List Reimbursements","operationId":"list_reimbursements_api_governance_reimbursements_get","parameters":[{"required":false,"schema":{"title":"Status","type":"string"},"name":"status","in":"query"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}},"post":{"tags":["governance"],"summary":"Create Reimbursement","operationId":"create_reimbursement_api_governance_reimbursements_post","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"title":"Body","type":"object"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/governance/reimbursements/{rid}/approve":{"post":{"tags":["governance"],"summary":"Approve Reimbursement","description":"三级流转：平台长→数字化复核→财务→已完成；任意级驳回即终止。\n\n分权规则：\n- 第 1 级（平台长审批）：tier ∈ {coach, backbone, boss}\n- 第 2 级（数字化复核）：仅 coach（数字化平台长）\n- 第 3 级（财务报销）：仅财务部 backbone 或 boss\n- 同一人不能审批同一单的连续两级（含已批准过的任一级）\n每级审批意见累加进 comment 列并写 audits。","operationId":"approve_reimbursement_api_governance_reimbursements__rid__approve_post","parameters":[{"required":true,"schema":{"title":"Rid","type":"integer"},"name":"rid","in":"path"},{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"requestBody":{"content":{"application/json":{"schema":{"title":"Body","type":"object"}}},"required":true},"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/governance/audits":{"get":{"tags":["governance"],"summary":"List Audits","operationId":"list_audits_api_governance_audits_get","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/governance/redlines":{"get":{"tags":["governance"],"summary":"Redlines","operationId":"redlines_api_governance_redlines_get","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/api/roadmap":{"get":{"tags":["roadmap"],"summary":"Roadmap","operationId":"roadmap_api_roadmap_get","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}},"/":{"get":{"summary":"Index","description":"GET /：index.html 存在则返回，否则兜底 JSON","operationId":"index__get","responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}}}}},"/api/heartbeat/run":{"post":{"summary":"Heartbeat Run","description":"立即执行一次心跳（日报 + 催办），返回执行摘要","operationId":"heartbeat_run_api_heartbeat_run_post","parameters":[{"required":false,"schema":{"title":"Authorization","type":"string"},"name":"authorization","in":"header"}],"responses":{"200":{"description":"Successful Response","content":{"application/json":{"schema":{}}}},"422":{"description":"Validation Error","content":{"application/json":{"schema":{"$ref":"#/components/schemas/HTTPValidationError"}}}}}}}},"components":{"schemas":{"HTTPValidationError":{"title":"HTTPValidationError","type":"object","properties":{"detail":{"title":"Detail","type":"array","items":{"$ref":"#/components/schemas/ValidationError"}}}},"LoginIn":{"title":"LoginIn","required":["person_id"],"type":"object","properties":{"person_id":{"title":"Person Id","type":"integer"}}},"ValidationError":{"title":"ValidationError","required":["loc","msg","type"],"type":"object","properties":{"loc":{"title":"Location","type":"array","items":{"anyOf":[{"type":"string"},{"type":"integer"}]}},"msg":{"title":"Message","type":"string"},"type":{"title":"Error Type","type":"string"}}}}}}